# docs/03-Design/OpenAPI/openapi.yaml
openapi: 3.0.3
info:
  title: Project API
  version: 0.1.0
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
...
paths:
...
  /api/v1/auth/end-user/register/request-otp:
    post:
      summary: Start end-user registration (request OTP)
      description: >
        Starts registration for an end user by validating input and sending an OTP via NOTIFICATIONS.
        Rate limited per phone number.
      tags:
        - AUTH
      operationId: Auth_EndUserRegister_RequestOtp
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndUserRegisterStartRequest'
      responses:
        '202':
          description: OTP requested successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
        '400':
          $ref: '#/components/responses/DefaultError'
        '409':
          $ref: '#/components/responses/DefaultError'
        '429':
          $ref: '#/components/responses/DefaultError'
  /api/v1/auth/end-user/register/verify-otp:
    post:
      summary: Verify OTP and complete end-user registration
      description: >
        Verifies the OTP for the given phone number and, on success, creates the end-user account and issues tokens.
      tags:
        - AUTH
      operationId: Auth_EndUserRegister_VerifyOtp
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndUserRegisterVerifyRequest'
      responses:
        '200':
          description: End-user registered and tokens issued
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/DefaultError'
        '401':
          $ref: '#/components/responses/DefaultError'
        '409':
          $ref: '#/components/responses/DefaultError'
        '429':
          $ref: '#/components/responses/DefaultError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    CorrelationId:
      description: Correlation id for tracing
      schema:
        type: string
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Idempotency key for safely retrying POST requests with side effects.
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            correlationId:
              type: string
      required:
        - error
...
    LoginResponse:
      description: Standard auth tokens response for both end-user and admin logins.
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: seconds
        tokenType:
          type: string
          description: Token type for the Authorization header.
          default: Bearer
        user:
          $ref: '#/components/schemas/EndUserProfileSummary'
        admin:
          $ref: '#/components/schemas/AdminProfileSummary'
        roles:
          type: array
          description: Roles granted to the authenticated principal.
          items:
            type: string
      required:
        - accessToken
        - refreshToken
        - expiresIn
    EndUserProfileSummary:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        fullName:
          type: string
        phone:
          type: string
        companyId:
          type: integer
          format: int64
        tier:
          type: integer
          description: 1/2/3 according to USERS module.
      required:
        - userId
        - fullName
        - phone
        - companyId
        - tier
    AdminProfileSummary:
      type: object
      properties:
        adminUserId:
          type: integer
          format: int64
        loginId:
          type: string
      required:
        - adminUserId
        - loginId
    DeviceInfo:
      type: object
      description: Optional device/browser fingerprint used to bind tokens to a device.
      properties:
        deviceId:
          type: string
          maxLength: 128
        userAgent:
          type: string
          maxLength: 512
        platform:
          type: string
          maxLength: 128
    EndUserRegisterStartRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 200
        companyId:
          type: integer
          format: int64
        phone:
          type: string
          maxLength: 20
        acceptTerms:
          type: boolean
      required:
        - fullName
        - companyId
        - phone
        - acceptTerms
    EndUserRegisterVerifyRequest:
      type: object
      properties:
        phone:
          type: string
          maxLength: 20
        otpCode:
          type: string
          maxLength: 10
        deviceInfo:
          $ref: '#/components/schemas/DeviceInfo'
      required:
        - phone
        - otpCode
...
  responses:
    DefaultError:
      description: Error
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'